Move available IP/prefix endpoints to API view mixins
